Update 03_booleans_operators_date.md
The output "undefined" you see is the default result of executing the code. It indicates that there is no specific value to return or display after running the provided code snippets.

Also note that the output "undefined" you see is the default result of executing the code. It indicates that there is no specific value to return or display after running the provided code snippets.